Bug 578073 - [18] JSR 269 - Implement Elements.getFileObjectOf(Element)
Incorporate review comments from Stephan
