+++ org.eclipse.jdt.compiler.apt.tests/resources/mod_locations/modules/mod.a/abc/internal/A.java
@@ -4,4 +4,5 @@ public class A extends TypeInAModule {
 	public void aMethod() {
 		// Nothing here
 	}
+	class InnerTypeInAModule {}
 }
